LPS-185680 Extract TreeViewComponent to its own file - Get selectedObjectFolder name through context instead of passing it as a parameter - Rename lfr-objects__model-builder-left-sidebar-btn-folder class to lfr-objects__model-builder-left-sidebar-go-to-folder-button
