+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ModelBuilder/EditObjectFolder.tsx
@@ -241,10 +241,7 @@ export default function EditObjectFolder({
 				setShowModal={setShowModal}
 			/>
 			<div className="lfr-objects__model-builder-diagram-container">
-				<LeftSidebar
-					selectedObjectFolderName={selectedObjectFolder.name}
-					setShowModal={setShowModal}
-				/>
+				<LeftSidebar setShowModal={setShowModal} />
 
 				<Diagram setShowModal={setShowModal} />
