+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ModelBuilder/LeftSidebar/TreeViewComponent.d.ts
@@ -1,0 +1,15 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+/// <reference types="react" />
+
+import './LeftSidebar.scss';
+export default function TreeViewComponent({
+	query,
+	showActions,
+}: {
+	query: string;
+	showActions?: boolean;
+}): JSX.Element;
